chore(src): reorganize Toast elements
