+++ src/helix-ui/elements/hx-toast/_shadow.less
@@ -1,4 +1,4 @@
-@import "./HXElement";
+@import "../HXElement";
 @import (reference) 'components/box';
 @import (reference) 'components/buttons';
