+++ src/helix-ui/elements/hx-toast/index.js
@@ -1,6 +1,7 @@
-import { HXElement } from './HXElement';
-import shadowMarkup from './HXToastElement.html';
-import shadowStyles from './HXToastElement.less';
+import { HXElement } from '../HXElement.js';
+
+import shadowMarkup from './_shadow.html';
+import shadowStyles from './_shadow.less';
 
 const ICONS = {
     'error': 'exclamation-circle',
